% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_re_index.R
\name{check_and_re_index}
\alias{check_and_re_index}
\title{Re-indexes list of Hafvog data if necessary}
\usage{
check_and_re_index(surv_list)
}
\arguments{
\item{surv_list}{List of Hafog data read infrom multiple survey zips}
}
\value{
Original list if no overlap, otherwise re-indexed version
}
\description{
Re-indexes list of Hafvog data if necessary
}
