% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_fix_station_overlap.R
\name{check_and_re_number_stations}
\alias{check_and_re_number_stations}
\title{Re-number stations with jump between surveys in alist of Hafvog data if necessary}
\usage{
check_and_re_number_stations(surv_list, jump_by = 100)
}
\arguments{
\item{surv_list}{List of Hafog data read in from multiple survey zips}
}
\value{
Original list if no station nummber overlap, otherwise re-numbered stations version
}
\description{
Re-number stations with jump between surveys in alist of Hafvog data if necessary
}
