% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_timeperiods.R
\name{hv_bind_timeperiods}
\alias{hv_bind_timeperiods}
\title{Title}
\usage{
hv_bind_timeperiods(current, historical)
}
\arguments{
\item{current}{a list}

\item{historical}{a list}
}
\value{
a list
}
\description{
Title
}
