% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_fix_station_overlap.R,
%   R/check_and_re_index.R
\name{hv_concat}
\alias{hv_concat}
\title{Concatenates a list of Hafvog data}
\usage{
hv_concat(surv_list)

hv_concat(surv_list)
}
\arguments{
\item{surv_list}{List of Hafog data read infrom multiple survey zips}
}
\value{
A single hafvog data list with concatenated
frames 'ST','M','KV','LE','pred','prey'.

A single hafvog data list with concatenated
frames 'ST','M','KV','LE','pred','prey'.
}
\description{
Concatenates a list of Hafvog data

Concatenates a list of Hafvog data
}
