% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{hv_create_tables}
\alias{hv_create_tables}
\title{Create tables from hafvog "skraning"}
\usage{
hv_create_tables(list, scale = TRUE)
}
\arguments{
\item{list}{A list object containing hafvog dataframes}

\item{scale}{A boolean (default TRUE) will raise number of length measured
with that counted}
}
\value{
a list
}
\description{
Normally the input to this function is what is read in from hv_import.
The outputs are tables like fiskar.lengdir, fiskar.kvarnir, fiskar.numer, ...
}
