% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{hv_import_zipfile}
\alias{hv_import_zipfile}
\title{Read any hafvog zipfile}
\usage{
hv_import_zipfile(zipfile)
}
\arguments{
\item{zipfile}{The pathname of the zip file}
}
\value{
A list of tibbles
}
\description{
The files to read from within the zipfile are all jsonfiles
}
