% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hv_measures}
\alias{hv_measures}
\title{The measurement (skraning) table}
\usage{
hv_measures(pth, std = TRUE)
}
\arguments{
\item{pth}{A character string specifying the path to to the data}

\item{std}{Boolean (default TRUE) specifying if names should be standardized}
}
\value{
A tibble
}
\description{
The measurement (skraning) table
}
