% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zipfiles.R
\name{hv_read_hafvog}
\alias{hv_read_hafvog}
\title{Reads muliple hafvogs zipfiles}
\usage{
hv_read_hafvog(zipfiles, collapse_station = TRUE)
}
\arguments{
\item{collapse_station}{boolean (default TRUE), returns station, towstation and environment as a single table}

\item{zip_file}{File name, including path}
}
\value{
a list
}
\description{
Reads muliple hafvogs zipfiles
}
