% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{hv_read_historical}
\alias{hv_read_historical}
\title{Read historical data}
\usage{
hv_read_historical(AR, SYNAFLOKKUR)
}
\arguments{
\item{AR}{The current data year, year less than that will be returned}

\item{SYNAFLOKKUR}{The survey sample class, normally either 30 or 35}
}
\value{
a list
}
\description{
Read historical data
}
