% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_tables.R
\name{hv_read_stillingar}
\alias{hv_read_stillingar}
\title{Title}
\usage{
hv_read_stillingar(
  zipfile = "data-raw/stillingar/stillingar_SMB_rall_(botnfiskur).zip"
)
}
\arguments{
\item{zipfile}{The pathname of the zip file}
}
\description{
The current form is just mimicry
}
