% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{hv_read_stodtoflur}
\alias{hv_read_stodtoflur}
\title{Reads hafvog's 'stodtoflur'}
\usage{
hv_read_stodtoflur(zipfile)
}
\arguments{
\item{zipfile}{The pathname of the zip file}
}
\value{
A list of tibbles
}
\description{
Reads hafvog's 'stodtoflur'
}
