% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zip_data.R
\name{hv_read_zip_data}
\alias{hv_read_zip_data}
\title{Reads muliple hafvogs files}
\usage{
hv_read_zip_data(zip_file)
}
\arguments{
\item{zip_file}{File name, including path}
}
\value{
a list
}
\description{
Reads muliple hafvogs files
}
