% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{hv_read_zipfile}
\alias{hv_read_zipfile}
\title{Read a single hafvog zipfile}
\usage{
hv_read_zipfile(zipfile)
}
\arguments{
\item{zipfile}{The name, including path of the hafvog zipfile}
}
\value{
a list containing individual tables
}
\description{
Read a single hafvog zipfile
}
