% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{hv_tbl}
\alias{hv_tbl}
\title{The base table function}
\usage{
hv_tbl(pth, tbl)
}
\arguments{
\item{pth}{A character string specifying the path to to the data}

\item{tbl}{The actual table name}
}
\value{
A tibble
}
\description{
The base table function
}
