% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stillingar.R
\name{hv_tidy_length_weights}
\alias{hv_tidy_length_weights}
\title{Create a tidy QC species length-weight table}
\usage{
hv_tidy_length_weights(stillingar, adjust_lengths = TRUE)
}
\arguments{
\item{stillingar}{A tibble}

\item{adjust_lengths}{A boolean (default TRUE) where the max length for
the length-weight is potentially obtained from the "range" table. And
if only one value for length-weights, gets the minimum from the range table}
}
\value{
A tibble
}
\description{
The resulting tibble contains ...
}
\details{
The datasource is a list that contains object "fiskteg_lengd_thyngd". The list
is normally obtained using hv_read_stillingar.
}
