% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stillingar.R
\name{hv_tidy_range}
\alias{hv_tidy_range}
\title{Create a tidy QC species min-max table}
\usage{
hv_tidy_range(stillingar, long = TRUE)
}
\arguments{
\item{stillingar}{A tibble}

\item{long}{Boolean (default TRUE) returns a long table}
}
\value{
A tibble
}
\description{
The resulting tibble contains the min and max lengd for a species and the
ratio range of values relative to the ungutted species weight.
}
\details{
The datasource is a list that contains object "fiskteg_tegundir". The list
is normally obtained using hv_read_stillingar.
}
\note{
The variable "maelibretti" is the odd man out
}
