% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hv_towstations}
\alias{hv_towstations}
\title{The towstation (togstodvar) table}
\usage{
hv_towstations(pth, std = TRUE)
}
\arguments{
\item{pth}{A character string specifying the path to to the data}

\item{std}{Boolean (default TRUE) specifying if names should be standardized}
}
\value{
A tibble
}
\description{
The towstation (togstodvar) table
}
