% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_re_index.R
\name{id_overlap}
\alias{id_overlap}
\title{Checks for index overlap in a list of multiple hafvog survey files}
\usage{
id_overlap(surv_list)
}
\arguments{
\item{zip_file}{List of hafog inputs read infrom multiple survey zips}
}
\value{
logical, is there overlap in the 'synis_id'-s from different hafvog instances
}
\description{
Checks for index overlap in a list of multiple hafvog survey files
}
