% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_tables.R
\name{js_stadla_rallstodvar}
\alias{js_stadla_rallstodvar}
\title{Title}
\usage{
js_stadla_rallstodvar(con = "data-raw/stillingar_SMB_rall_(botnfiskur).zip")
}
\arguments{
\item{con}{path to stillingar_SMB_rall}
}
\description{
The current form is just mimicry
}
