% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_tables.R
\name{js_stadla_tegund_smb}
\alias{js_stadla_tegund_smb}
\title{Title}
\usage{
js_stadla_tegund_smb(con = "inst/extdata/stillingar_SMB_rall_(botnfiskur).zip")
}
\arguments{
\item{con}{XX}
}
\description{
bla, bla
}
