% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_tables.R
\name{js_stillingar_all}
\alias{js_stillingar_all}
\title{Title}
\usage{
js_stillingar_all(
  con = "data-raw/stillingar/stillingar_SMB_rall_(botnfiskur).zip"
)
}
\arguments{
\item{con}{path to stillingar_SMB_rall}
}
\description{
The current form is just mimicry
}
