% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_re_index.R
\name{new_index}
\alias{new_index}
\title{Generates non-overlapping indices for a list of hafvog surveys}
\usage{
new_index(surv_list)
}
\arguments{
\item{surv_list}{List of Hafog data read infrom multiple survey zips}
}
\value{
list of sequential id-s one after the other for the multiple inputs
}
\description{
Generates non-overlapping indices for a list of hafvog surveys
}
