% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_fix_station_overlap.R
\name{new_stations}
\alias{new_stations}
\title{Generates non-overlapping indices for a list of hafvog surveys}
\usage{
new_stations(surv_list, jump_by = 100)
}
\arguments{
\item{surv_list}{List of Hafog data read infrom multiple survey zips}
}
\value{
list of sequential id-s one after the other for the multiple inputs
}
\description{
Generates non-overlapping indices for a list of hafvog surveys
}
