% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_re_index.R
\name{re_index}
\alias{re_index}
\title{Re-indexes the second to last table of a survey input with a new index}
\usage{
re_index(surv)
}
\arguments{
\item{surv}{Hafog data read in from single survey}
}
\value{
survey list, column 'newid' added to all tables
}
\description{
Re-indexes the second to last table of a survey input with a new index
}
