% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_fix_station_overlap.R
\name{st_overlap}
\alias{st_overlap}
\title{Checks for station number overlap in a list of multiple hafvog survey files}
\usage{
st_overlap(surv_list)
}
\arguments{
\item{zip_file}{List of hafog inputs read infrom multiple survey zips}
}
\value{
logical, is there overlap in the 'stod'/stations numbers from different hafvog instance/surveys/vessels
}
\description{
Checks for station number overlap in a list of multiple hafvog survey files
}
