% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_tables.R
\name{tbl_json}
\alias{tbl_json}
\title{Read in a single json file from a zip-file}
\usage{
tbl_json(zipfile, file)
}
\arguments{
\item{zipfile}{The path and the name of the zip file}

\item{file}{The json file to read}
}
\value{
a tibble
}
\description{
Read in a single json file from a zip-file
}
