% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vocabulary}
\alias{vocabulary}
\title{Variable vocabulary in {fmr}}
\format{
\subsection{\code{vocabulary}}{

A vector
}
}
\usage{
vocabulary
}
\description{
A vector containing variable names as delivered in the json-files and
the corresponding names used in {hafvog} if argument std is set to TRUE
}
\keyword{datasets}
