% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{hv_read_hafvog}
\alias{hv_read_hafvog}
\title{Reads hafvog's zipfiles}
\usage{
hv_read_hafvog(zipfiles, collapse_station = TRUE)
}
\arguments{
\item{zipfiles}{File names, including path}

\item{collapse_station}{boolean (default TRUE), returns station, towstation and environment as a single table}
}
\value{
a list
}
\description{
Can read in multiple files. Name of source file is stored in variable ".id"
}
