% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_historical_data.R
\name{hv_read_historical}
\alias{hv_read_historical}
\title{Read historical data}
\usage{
hv_read_historical(years, sample_class)
}
\arguments{
\item{years}{Years of data to be returned}

\item{sample_class}{The survey sample class, normally either 30 or 35}
}
\value{
a list
}
\description{
Read historical data
}
