% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{hv_read_stillingar}
\alias{hv_read_stillingar}
\title{Reads hafvog's 'stillingar'}
\usage{
hv_read_stillingar(zipfile)
}
\arguments{
\item{zipfile}{The pathname of the zip file}
}
\value{
A list of tibbles
}
\description{
Reads hafvog's 'stillingar'
}
