% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zip_data.R
\name{hv_read_zip_file}
\alias{hv_read_zip_file}
\title{Reads a sinle hafvog zip file}
\usage{
hv_read_zip_file(zip_file)
}
\arguments{
\item{zip_file}{File name, including path}
}
\value{
a list
}
\description{
Reads a sinle hafvog zip file
}
