% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hv_station}
\alias{hv_station}
\title{The station (stodvar) table}
\usage{
hv_station(pth, std = TRUE, trim = TRUE)
}
\arguments{
\item{pth}{A character string specifying the path to to the data}

\item{std}{Boolean (default TRUE) specifying if names should be standardized}

\item{trim}{Boolean (default TRUE) to limit number of variables returned}
}
\value{
A tibble
}
\description{
The station (stodvar) table
}
